Move ResourceFingerprintVisitor to fhir-model
And the corresponding SaltHash class to `com.ibm.fhir.model.util`.
Per suggestion during review.
