+++ fhir-persistence-jdbc/src/test/java/com/ibm/fhir/persistence/jdbc/test/spec/BaseOperation.java
@@ -7,10 +7,10 @@
 package com.ibm.fhir.persistence.jdbc.test.spec;
 
 import com.ibm.fhir.model.resource.Resource;
+import com.ibm.fhir.model.util.SaltHash;
 import com.ibm.fhir.model.util.test.ResourceComparatorVisitor;
+import com.ibm.fhir.model.visitor.ResourceFingerprintVisitor;
 import com.ibm.fhir.persistence.exception.FHIRPersistenceException;
-import com.ibm.fhir.persistence.util.ResourceFingerprintVisitor;
-import com.ibm.fhir.persistence.util.SaltHash;
 
 /**
  * Create the resource using the persistence
