+++ fhir-persistence-jdbc/src/test/java/com/ibm/fhir/persistence/jdbc/test/spec/R4JDBCExamplesProcessor.java
@@ -19,13 +19,13 @@ import com.ibm.fhir.database.utils.api.ITransaction;
 import com.ibm.fhir.database.utils.api.ITransactionProvider;
 import com.ibm.fhir.model.resource.Resource;
 import com.ibm.fhir.model.spec.test.IExampleProcessor;
+import com.ibm.fhir.model.visitor.ResourceFingerprintVisitor;
 import com.ibm.fhir.persistence.FHIRPersistence;
 import com.ibm.fhir.persistence.context.FHIRHistoryContext;
 import com.ibm.fhir.persistence.context.FHIRPersistenceContext;
 import com.ibm.fhir.persistence.context.FHIRPersistenceContextFactory;
 import com.ibm.fhir.persistence.jdbc.FHIRPersistenceJDBCCache;
 import com.ibm.fhir.persistence.jdbc.impl.FHIRPersistenceJDBCImpl;
-import com.ibm.fhir.persistence.util.ResourceFingerprintVisitor;
 
 /**
  * Reads R4 example resources and performs a sequence of persistance
