+++ fhir-server-test/src/test/java/com/ibm/fhir/server/test/examples/ExampleRequestProcessor.java
@@ -19,9 +19,9 @@ import com.ibm.fhir.core.FHIRMediaType;
 import com.ibm.fhir.model.resource.Resource;
 import com.ibm.fhir.model.spec.test.DriverMetrics;
 import com.ibm.fhir.model.spec.test.IExampleProcessor;
+import com.ibm.fhir.model.util.SaltHash;
 import com.ibm.fhir.model.util.test.ResourceComparatorVisitor;
-import com.ibm.fhir.persistence.util.ResourceFingerprintVisitor;
-import com.ibm.fhir.persistence.util.SaltHash;
+import com.ibm.fhir.model.visitor.ResourceFingerprintVisitor;
 import com.ibm.fhir.server.test.FHIRServerTestBase;
 
 /**
